feat: Avaliando motorista

+++ src/main/java/com/carona/AppDeCarona/controller/dto/Usuario/DetalharUsuarioDto.java
@@ -5,7 +5,7 @@ import com.carona.AppDeCarona.entity.Usuario;
 public record DetalharUsuarioDto(String nome,
                                  String email,
                                  String password,
-                                 int avaliacao) {
+                                 double avaliacao) {
     public DetalharUsuarioDto(Usuario usuario){
         this(usuario.getNome(), usuario.getEmail(), usuario.getPassword(), usuario.getAvaliacao());
     }
